requirements.txt
pytest
redis
loguru
pandas
pytz
sqlalchemy
psycopg2-binary
gdown
flask
flask_cors
python-dotenv
black
flake8
isort